+++ src/main/java/lotto/model/LottoService.java
@@ -4,7 +4,7 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import lotto.configurations.LottoConstants;
+import lotto.configuration.LottoConstants;
 
 public class LottoService {
